all order and create order and long bd and int[] bd working

+++ Typography_course_work/src/main/java/com/example/Typography_course_work/model/Product.java
@@ -25,7 +25,7 @@ public class Product {
     private String title;
 
     @Column(name = "materials")
-    private List<Long> materials; // Для связи через массив id
+    private List<Integer> materials; // Для связи через массив id
 
     @Column(nullable = false)
     private Integer price;
